tweak: generalize BackgroundWorker
